fastapi/pyproject.toml
[project]
name = "todos"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)"
]

[tool.poetry.scripts]
dev = "todos.main:start"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
